+++ src/widgets/ui/Header/ProfileModal/api/useNetworkLogout.ts
@@ -1,17 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
-import axios from 'axios';
 
 import URL from 'shared/config/urls';
-
-// axios 인스턴스 생성
-const axiosInstance = axios.create({
-  baseURL: URL.BASEURL,
-  withCredentials: true,
-  timeout: 5000, // 임의로 지정
-  headers: {
-    'Content-Type': 'application/json',
-  },
-});
+import axiosInstance from 'shared/hooks/useAxiosInstance';
 
 // 메인데이터 호출
 const fetchLogout = async () => {
